neopixel into abstract parts, clean up swm
