+++ edg/parts/human_interface/Neopixel.py
@@ -6,20 +6,6 @@ from ...circuits import *
 from ...vendor_parts.jlc.JlcPart import JlcPart
 
 
-@abstract_block_default(lambda: Ws2812b)
-class Neopixel(Light, Block):
-    """Abstract base class for individually-addressable, serially-connected Neopixel-type
-    (typically RGB) LEDs and defines the pwr/gnd/din/dout interface."""
-
-    def __init__(self) -> None:
-        super().__init__()
-        self.pwr = self.Port(VoltageSink.empty(), [Power])
-        self.vdd = self.pwr  # deprecated alias
-        self.gnd = self.Port(Ground.empty(), [Common])
-        self.din = self.Port(DigitalSink.empty(), [Input])
-        self.dout = self.Port(DigitalSource.empty(), optional=True)
-
-
 class Ws2812b(Neopixel, FootprintBlock, JlcPart):
     """5050-size Neopixel RGB. Specifically does NOT need extra filtering capacitors."""
